Nasiko/talentflowHR/requirements.txt
fastapi>=0.109.0
uvicorn>=0.27.0
pydantic>=2.6.0
python-dotenv>=1.0.0
requests>=2.31.0
numpy>=2.0.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-google-genai>=4.0.0
google-generativeai>=0.8.0
click>=8.1.7
motor>=3.3.0
pymongo>=4.6.0
python-multipart>=0.0.9
pdfplumber>=0.10.0
pymupdf>=1.24.0
google-api-python-client>=2.120.0
google-auth>=2.28.0
google-auth-oauthlib>=1.2.0
